#6 ファイル移動の実装 #5 重複ファイルの抽出 再生時間が短い動画のサムネイル生成変更 サムネイル再生成機能追加

+++ XMovie/Models/Thumbnail.cs
@@ -19,7 +19,7 @@ namespace XMovie.Models
         [Required]
         public string FileName { get; set; }
 
-        public int Seconds { get; set; }
+        public double Seconds { get; set; }
 
         [ForeignKey("MovieId")]
         public virtual Movie Movie { get; set; }
